+++ src/app/(app)/@quests/speed-dial.tsx
@@ -9,9 +9,7 @@ import {
   BookUserIcon,
   BoxIcon,
   CircleXIcon,
-  ImportIcon,
   LogOutIcon,
-  UploadIcon,
 } from "lucide-react";
 import { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
 import { ReadonlyURLSearchParams, useRouter, useSearchParams } from "next/navigation";
@@ -19,15 +17,12 @@ import { ReadonlyURLSearchParams, useRouter, useSearchParams } from "next/naviga
 import { truncate_sessions } from "@/actions/logout";
 import { ModalParam } from "@/app/(app)/@modals/params";
 import Speeddial, { SpeeddialDirection } from "@/components/animata/container/speed-dial";
-import dataCommands from "@/components/header/dev-console/commands/data";
-import { TerminalContext } from "@/components/ui/terminal";
 import { updateParam } from "@/lib/params";
 
 export default function App(): React.ReactElement {
   const router: AppRouterInstance = useRouter();
   const searchParams: ReadonlyURLSearchParams = useSearchParams();
 
-  // TODO(ayvi): unhide export/import functions
   const actionButtons = [
     {
       icon: <BookUserIcon size={14} stroke="#f0f0ff" />,
@@ -49,22 +44,6 @@ export default function App(): React.ReactElement {
       key: "add-quest",
       buttonAction: () =>
         updateParam(router, searchParams, [{ key: "modal", value: ModalParam.AddQuest }]),
-    },
-    {
-      icon: <UploadIcon size={14} stroke="#f0f0ff" />,
-      label: <p className="text-xs">Export Data</p>,
-      key: "export-data",
-      hidden: true,
-      buttonAction: async () =>
-        await dataCommands[0].run({ args: [] } as unknown as TerminalContext),
-    },
-    {
-      icon: <ImportIcon size={14} stroke="#f0f0ff" />,
-      label: <p className="text-xs">Import Data</p>,
-      key: "import-data",
-      hidden: true,
-      buttonAction: async () =>
-        await dataCommands[1].run({ args: [] } as unknown as TerminalContext),
     },
     {
       icon: <LogOutIcon size={14} stroke="#f0f0ff" />,